requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-dotenv==1.0.0
pandas==2.1.2
numpy==1.26.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
python-dateutil==2.8.2
requests==2.31.0
pytest==7.4.4
httpx==0.25.2
